simplify CAS with educonnect #56
