+++ pronotepy/ent/__init__.py
@@ -1,4 +1,5 @@
 from .simple_auth import *
 from .educonnect import *
+from .cas_edu import *
 from .open_ent_ng import *
 from .open_ent_ng_educonnect import *
