+++ pronotepy/ent/educonnect.py
@@ -102,160 +102,6 @@ def ac_rennes(username: str, password: str) -> requests.cookies.RequestsCookieJa
     return session.cookies
 
 
-def occitanie_montpellier_educonnect(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for Occitanie Montpellier
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://cas.mon-ent-occitanie.fr/login'
-    # ENT Connection
-    session = requests.Session()
-
-    params = {
-        'selection': 'MONT-EDU_parent_eleve',
-        'submit': 'Valider'
-    }
-
-    response = session.get(ent_login, params=params, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"]
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    log.debug(f'[ENT Occitanie] Logging in with {username}')
-
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def ent_auvergnerhonealpe(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for Auvergne Rhone Alpes with Educonnect
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://cas.ent.auvergnerhonealpes.fr/login?selection=EDU&submit=Valider'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def ac_orleans_tours(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for AC Orleans-Tours
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://ent.netocentre.fr/cas/login?&idpId=parentEleveEN-IdP'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def monbureaunumerique(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for MonBureauNumerique (Grand Est)
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://cas.monbureaunumerique.fr/login?selection=EDU'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"],
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def ac_reims(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for AC Reims
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    return monbureaunumerique(username, password)
-
-
 def ent_elyco(username: str, password: str) -> requests.cookies.RequestsCookieJar:
     ent_login_page = 'https://cas3.e-lyco.fr/discovery/WAYF'
     session = requests.Session()
@@ -269,85 +115,7 @@ def ent_elyco(username: str, password: str) -> requests.cookies.RequestsCookieJa
     }
 
     response = session.get(ent_login_page, params=params, headers=HEADERS)
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def eclat_bfc(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for Eclat BFC
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login = 'https://cas.eclat-bfc.fr/login'
-    # ENT Connection
-    session = requests.Session()
-
-    params = {
-        'selection': 'EDU',
-        'submit': 'Valider'
-    }
-
-    response = session.get(ent_login, params=params, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"]
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    log.debug(f'[ENT Eclat BFC] Logging in with {username}')
-
-    educonnect(response.url, session, username, password)
-
-    return session.cookies
-
-
-def cas_agora06_educonnect(username: str, password: str) -> requests.cookies.RequestsCookieJar:
-    """
-    ENT for Agora06 with Educonnect
-
-    Parameters
-    ----------
-    username : str
-        username
-    password : str
-        password
-
-    Returns
-    -------
-    cookies : cookies
-        returns the ent session cookies
-    """
-    # ENT / PRONOTE required URLs
-    ent_login_page = 'https://cas.agora06.fr/login?selection=EDU&submit=Valider'
-
-    # ENT Connection
-    session = requests.Session()
-
-    # Connection URL specifying the pronote service
-    response = session.get(ent_login_page, headers=HEADERS)
-    soup = BeautifulSoup(response.text, 'html.parser')
-    payload = {
-        "RelayState": soup.find("input", {"name": "RelayState"})["value"],
-        "SAMLRequest": soup.find("input", {"name": "SAMLRequest"})["value"]
-        }
-
-    response = session.post(soup.find("form")['action'], data=payload, headers=HEADERS)
-    log.debug(f'[CAS Agora 06] Logging in with {username}')
-
+    log.debug(f'[ENT Elyco] Logging in with {username}')
     educonnect(response.url, session, username, password)
 
     return session.cookies
